opsx/requirements.txt
Django==1.6.5
django-crispy-forms==1.4.0
django-reversion==1.8.1
django-xadmin==0.5.0
